requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pymongo==3.12.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
